Move result related types to scylla-cql-core
