+++ scylla-cql-core/src/frame/mod.rs
@@ -1,3 +1,4 @@
+pub mod response;
 pub mod types;
 
 use std::fmt::Display;
