+++ scylla-cql-core/src/frame/response/mod.rs
@@ -1,0 +1,3 @@
+//! CQL protocol-level representation of a `RESULT` response.
+
+pub mod result;
