+++ scylla/src/cluster/metadata/mod.rs
@@ -30,7 +30,7 @@ use thiserror::Error;
 use uuid::Uuid;
 
 // Re-export of CQL types.
-pub use scylla_cql::frame::response::result::{
+pub use scylla_cql_core::frame::response::result::{
     CollectionType, ColumnType, NativeType, UserDefinedType,
 };
